Move viscera to flesh; viscera is for something else :)
