+++ alburnum/maas/tests/test_flesh.py
@@ -12,7 +12,7 @@ from textwrap import dedent
 
 from alburnum.maas import (
     auth,
-    viscera,
+    flesh,
 )
 from alburnum.maas.testing import (
     make_name,
@@ -24,11 +24,11 @@ from alburnum.maas.tests.test_auth import make_options
 class TestLogin(TestCase):
 
     def test_cmd_login_ensures_valid_apikey(self):
-        parser = viscera.ArgumentParser()
+        parser = flesh.ArgumentParser()
         options = make_options()
-        check_key = self.patch(viscera, "check_valid_apikey")
+        check_key = self.patch(flesh, "check_valid_apikey")
         check_key.return_value = False
-        login = viscera.cmd_login(parser)
+        login = flesh.cmd_login(parser)
         error = self.assertRaises(SystemExit, login, options)
         self.assertEqual(
             "The MAAS server rejected your API key.",
@@ -40,7 +40,7 @@ class TestLogin(TestCase):
     def test_print_whats_next(self):
         profile = {"name": make_name("profile"), "url": make_name("url")}
         stdout = self.patch(sys, "stdout", StringIO())
-        viscera.cmd_login.print_whats_next(profile)
+        flesh.cmd_login.print_whats_next(profile)
         expected = dedent("""\
 
             You are now logged in to the MAAS server at %(url)s with the
